Move example into remaining plugin
