+++ plugins/file_selector/file_selector_linux/example/analysis_options.yaml
@@ -1,0 +1,1 @@
+include: ../../../../analysis_options.yaml
