+++ plugins/file_selector/file_selector_linux/example/lib/main.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:example/home_page.dart';
-import 'package:example/get_directory_page.dart';
-import 'package:example/open_text_page.dart';
-import 'package:example/open_image_page.dart';
-import 'package:example/open_multiple_images_page.dart';
-import 'package:example/save_text_page.dart';
+
+import 'get_directory_page.dart';
+import 'home_page.dart';
+import 'open_text_page.dart';
+import 'open_image_page.dart';
+import 'open_multiple_images_page.dart';
+import 'save_text_page.dart';
 
 void main() {
   runApp(MyApp());
