+++ plugins/file_selector/file_selector_linux/example/lib/open_text_page.dart
