+++ plugins/file_selector/file_selector_linux/example/linux/.gitignore
